requirements.txt
uvicorn~=0.29.0
tensorflow~=2.16.2
numpy~=1.26.4
opencv-python~=4.10.0.84
fastapi~=0.111.0
pillow~=10.3.0